backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
